Enables editing and deleting of stems and child stems

+++ frontend/modules/create/containers/createStemsContainer.jsx
@@ -3,6 +3,7 @@ import CreateStems from '../components/createStems';
 import WithCache from '~/core/cache/containers/withCache';
 import axios from 'axios';
 import handleRequestError from '~/core/app/helpers/handleRequestError';
+import addModal from '~/core/dialogs/helpers/addModal';
 import getCache from '~/core/cache/helpers/getCache';
 import filter from 'lodash/filter';
 import WithRouter from '~/core/app/components/withRouter';
@@ -68,15 +69,66 @@ class CreateStemsContainer extends Component {
     }
   }
 
+  onEditStemClicked = (stem) => {
+    addModal({
+      title: 'Edit stem',
+      schema: {
+        name: {
+          type: 'Text',
+          label: 'Name',
+          shouldAutoFocus: true
+        },
+        description: {
+          type: 'TextArea',
+          label: 'Description',
+          features: ['bold', 'italic', 'underline', 'strikethrough', 'code', 'blockquote', 'link', 'leftAlign', 'centerAlign', 'rightAlign', 'justifyAlign', 'bulletedList', 'numberedList']
+        }
+      },
+      model: {
+        name: stem.name,
+        description: stem.description
+      },
+      actions: [{
+        type: 'CANCEL',
+        text: 'Cancel'
+      }, {
+        type: 'SAVE',
+        text: 'Save',
+        color: 'primary'
+      }]
+    }, (state, { type, modal }) => {
+      if (state === 'ACTION' && type === 'SAVE') {
+        axios.put(`/api/stems/${stem._id}`, modal).then(() => {
+          this.props.stems.fetch();
+        }).catch(handleRequestError);
+      }
+    });
+  }
+
   onDeleteStemClicked = (stemId) => {
-    this.setState({ deletingId: stemId });
-    axios.delete(`/api/stems/${stemId}`).then(() => {
-      this.props.stems.fetch().then(() => {
-        this.setState({ deletingId: null });
-      });
-    }).catch((error) => {
-      this.setState({ deletingId: null });
-      handleRequestError(error);
+    addModal({
+      title: 'Delete stem',
+      description: 'Are you sure you want to delete this stem? All slides in this stem and child stems will be removed.',
+      actions: [{
+        type: 'CANCEL',
+        text: 'Cancel'
+      }, {
+        type: 'DELETE',
+        text: 'Delete',
+        color: 'danger'
+      }]
+    }, (state, { type }) => {
+      if (state === 'ACTION' && type === 'DELETE') {
+        this.setState({ deletingId: stemId });
+        axios.delete(`/api/stems/${stemId}`).then(() => {
+          this.props.stems.fetch().then(() => {
+            this.setState({ deletingId: null });
+          });
+        }).catch((error) => {
+          this.setState({ deletingId: null });
+          handleRequestError(error);
+        });
+      }
     });
   }
 
@@ -90,6 +142,7 @@ class CreateStemsContainer extends Component {
         deletingId={deletingId}
         getSlideCountForStem={this.getSlideCountForStem}
         onCreateStemClicked={this.onCreateStemClicked}
+        onEditStemClicked={this.onEditStemClicked}
         onDeleteStemClicked={this.onDeleteStemClicked}
         onStemClicked={this.onStemClicked}
       />
